Refactor fetchText into global function and remove from PromptContext
